clean dead code
